requirements-inference.txt
# --- Core ML ---
numpy
pandas
scikit-learn
xgboost
joblib

# --- API ---
fastapi
uvicorn
pydantic

# --- Config ---
pyyaml
